+++ src/Trade360SDK.Common.Entities/Entities/Fixtures/UniformColor.cs
@@ -1,0 +1,11 @@
+namespace Trade360SDK.Common.Entities.Fixtures
+{
+    public class UniformColor
+    {
+        public string? Primary { get; set; }
+
+        public string? Number { get; set; }
+
+        public string? Outline { get; set; }
+    }
+}