+++ src/Trade360SDK.Common.Entities/Entities/OutrightLeague/OutrightLeagueFixture.cs
@@ -16,6 +16,8 @@ namespace Trade360SDK.Common.Entities.OutrightLeague
 
         public Location? Location { get; set; }
 
+        public DateTime? StartDate { get; set; }
+
         public DateTime LastUpdate { get; set; }
 
         public FixtureStatus Status { get; set; }